feat: [GM-18] [BE] Create email register endpoint (#14)
* Update doc.yml

* Update doc.yml

* Add EmailRegister endpoint

* Add EmailRegister endpoint

* Add test cases

* Optimize unit tests

* Update go.mod

+++ src/go/user/app/models/base.go
@@ -24,6 +24,6 @@ type UUIDModel struct {
 	orm.Timestamps
 }
 
-func (r UUIDModel) GetID() uint {
-	return uint(snowflakeInstance.Generate())
+func (r UUIDModel) GetID() uint64 {
+	return uint64(snowflakeInstance.Generate())
 }
